# .migration-backup/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /branding-drafts:
    post:
      operationId: upsertBrandingDraft
      tags: [branding-drafts]
      summary: Create or update a branding draft for a game
      description: Upserts by the combination of gameId + draftToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertBrandingDraftRequest"
      responses:
        "200":
          description: The saved branding draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BrandingDraft"
  /branding-drafts/{id}:
    get:
      operationId: getBrandingDraft
      tags: [branding-drafts]
      summary: Get a branding draft by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The branding draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BrandingDraft"
        "404":
          description: Draft not found
  /branding-drafts/{id}/finalize:
    post:
      operationId: finalizeBrandingDraft
      tags: [branding-drafts]
      summary: Finalize a branding draft and create a pending order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FinalizeBrandingDraftRequest"
      responses:
        "200":
          description: The created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Draft not found
...
components:
  schemas:
...
    UpsertBrandingDraftRequest:
      type: object
      properties:
        gameId:
          type: string
        draftToken:
          type: string
        primaryColor:
          type: string
        secondaryColor:
          type: string
        accentColor:
          type: string
        logoDataUrl:
          type: string
          nullable: true
        bgDataUrl:
          type: string
          nullable: true
        fontDataUrl:
          type: string
          nullable: true
        heading:
          type: string
        tagline:
          type: string
          nullable: true
      required:
        - gameId
        - draftToken
        - primaryColor
        - secondaryColor
        - accentColor
        - heading
    BrandingDraft:
      type: object
      properties:
        id:
          type: string
        gameId:
          type: string
        draftToken:
          type: string
        primaryColor:
          type: string
        secondaryColor:
          type: string
        accentColor:
          type: string
        logoDataUrl:
          type: string
          nullable: true
        bgDataUrl:
          type: string
          nullable: true
        fontDataUrl:
          type: string
          nullable: true
        heading:
          type: string
        tagline:
          type: string
          nullable: true
        status:
          type: string
          enum: [draft, finalized]
      required:
        - id
        - gameId
        - draftToken
        - primaryColor
        - secondaryColor
        - accentColor
        - heading
        - status
    FinalizeBrandingDraftRequest:
      type: object
      properties:
        contactEmail:
          type: string
          nullable: true
    Order:
      type: object
      properties:
        id:
          type: string
        brandingDraftId:
          type: string
        gameId:
          type: string
        status:
          type: string
          enum: [pending_payment]
        totalAmountCents:
          type: integer
        contactEmail:
          type: string
          nullable: true
      required:
        - id
        - brandingDraftId
        - gameId
        - status
        - totalAmountCents
...